Use clang not Darwin version to detect c++14 support
Since we could be using a newer version of clang even
on an older operating system, detect the clang version
directly rather than the OS version to determine whether
or not to force C++11 support (which breaks C++14 support,
needed for CGAL, with newer clang). This should fix
conda builds, which don't use the system clang.